move utility traits and types
In this commit:

- move `Collector`, `PeerAddrLookup`, `Recipient`, `UdpMessage` from
  the `packet` module into the `util` module.
- add helper types that implement the traits: `EmptyCollector`,
  `BlackholeCollector`, `KnownSocketAddr`.
- add a few tests
