+++ monad-raptorcast/src/packet/builder.rs
@@ -26,7 +26,7 @@ use rand::Rng;
 use super::{
     assembler::{self, build_header, AssembleMode, PacketLayout},
     assigner::{self, ChunkAssignment},
-    BuildError, ChunkAssigner, UdpMessage,
+    BuildError, ChunkAssigner,
 };
 use crate::{
     message::MAX_MESSAGE_SIZE,
@@ -34,7 +34,10 @@ use crate::{
         GroupId, MAX_MERKLE_TREE_DEPTH, MAX_NUM_PACKETS, MAX_REDUNDANCY, MAX_SEGMENT_LENGTH,
         MIN_CHUNK_LENGTH, MIN_MERKLE_TREE_DEPTH,
     },
-    util::{compute_app_message_hash, unix_ts_ms_now, BroadcastMode, BuildTarget, Redundancy},
+    util::{
+        compute_app_message_hash, unix_ts_ms_now, BroadcastMode, BuildTarget, Collector,
+        Redundancy, UdpMessage,
+    },
 };
 
 pub const DEFAULT_MERKLE_TREE_DEPTH: u8 = 6;
@@ -201,7 +204,7 @@ where
         collector: &mut C,
     ) -> Result<()>
     where
-        C: super::Collector<UdpMessage<CertificateSignaturePubKey<ST>>>,
+        C: Collector<UdpMessage<CertificateSignaturePubKey<ST>>>,
     {
         self.prepare()
             .build_into(app_message, build_target, collector)
@@ -394,7 +397,7 @@ where
         collector: &mut C,
     ) -> Result<()>
     where
-        C: super::Collector<super::UdpMessage<CertificateSignaturePubKey<ST>>>,
+        C: Collector<UdpMessage<CertificateSignaturePubKey<ST>>>,
     {
         // figure out the layout of the packet
         let segment_size = self.unwrap_segment_size()?;
